Use single modal for event CRUD and hide actions
